+++ controls/views.py
@@ -12,7 +12,6 @@ from urllib.parse import quote, urlunparse
 from uuid import uuid4
 
 import rtyaml
-import siteapp.views as project_nav
 import trestle.oscal.component as trestlecomponent
 import trestle.oscal.ssp as trestlessp
 from django.conf import settings
@@ -22,25 +21,37 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.db.models import Count, Q
 from django.db.models.functions import Lower
-from django.http import (Http404, HttpResponse, HttpResponseForbidden,
-                         HttpResponseNotAllowed, HttpResponseRedirect,
-                         JsonResponse)
+from django.http import (
+    Http404,
+    HttpResponse,
+    HttpResponseForbidden,
+    HttpResponseNotAllowed,
+    HttpResponseRedirect,
+    JsonResponse,
+)
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.utils.text import slugify
 from django.views import View
 from django.views.generic import ListView
 from simple_history.utils import update_change_reason
-from siteapp.models import Invitation, Organization, Project, Tag
-from siteapp.settings import GOVREADY_URL
-from system_settings.models import SystemSettings
 
 import controls.utils as utils
 from controls.oscal import Catalogs
+from siteapp.models import Invitation, Organization, Project, Tag
+from siteapp.settings import GOVREADY_URL
+from system_settings.models import SystemSettings
+from utils import project_navigation
 
-from .forms import (DeploymentForm, ElementEditForm, ElementForm,
-                    ImportOSCALComponentForm, PoamForm, StatementPoamForm,
-                    SystemAssessmentResultForm)
+from .forms import (
+    DeploymentForm,
+    ElementEditForm,
+    ElementForm,
+    ImportOSCALComponentForm,
+    PoamForm,
+    StatementPoamForm,
+    SystemAssessmentResultForm,
+)
 from .models import *
 from .utilities import *
 
@@ -208,7 +219,7 @@ def controls_selected(request, system_id):
             if catalog.catalog_key not in internal_catalog_keys
         ]
 
-        nav = project_nav.project_navigation(request, project)
+        nav = project_navigation(request, project)
 
         # Return the controls
         context = {
@@ -417,7 +428,7 @@ class SelectedComponentsList(ListView):
             context['project'] = project
             context['system'] = system
             context['elements'] = Element.objects.all().exclude(element_type='system')
-            context['nav'] = project_nav.project_navigation(self.request, project)
+            context['nav'] = project_navigation(self.request, project)
             context['send_invitation'] = Invitation.form_context_dict(self.request.user, project, [self.request.user]),
             return context
         else:
@@ -1167,7 +1178,7 @@ def project_component_editor(request, system_id, element_id, catalog_key=None, c
 
         cl_id = control_id if control_id else narrative.get('control_id')
         cat = get_catalog_data_by_control(catalog_key, cl_id)
-        nav = project_nav.project_navigation(request, project)
+        nav = project_navigation(request, project)
 
         # Return the system's element information
         context = {
@@ -1904,7 +1915,7 @@ def project_control_editor(request, system_id, catalog_key, cl_id, statement_id=
         }
 
         cat = get_catalog_data_by_control(catalog_key, cl_id)
-        nav = project_nav.project_navigation(request, project)
+        nav = project_navigation(request, project)
 
         context = {
             'catalog': cat,
@@ -3043,8 +3054,7 @@ def poam_export(request, system_id, format='xlsx'):
             from tempfile import NamedTemporaryFile
 
             from openpyxl import Workbook
-            from openpyxl.styles import (Alignment, Border, Font, PatternFill,
-                                         Side)
+            from openpyxl.styles import Alignment, Border, Font, PatternFill, Side
 
             wb = Workbook()
             ws = wb.active
