+++ siteapp/views.py
@@ -61,6 +61,8 @@ structlog.configure(processors=[structlog.processors.JSONRenderer()])
 logger = get_logger()
 # logger = logging.getLogger(__name__)
 
+from utils import project_navigation
+
 LOGIN = "login"
 SIGNUP = "signup"
 
@@ -1107,27 +1109,6 @@ def project(request, project):
         "percent_compliant": control_compliance_stats.get('percent_compliant'),
         "nav": nav,
     })
-
-def project_navigation(request, project):
-    task = project.root_task.get_or_create_subtask(request.user, "ssp_intro")
-    nav = {
-        "controls": {
-            "url": "/systems/" + str(project.system.id) + "/controls/selected",
-            "title": "Review Controls",
-            "id": "project-controls",
-        },
-        "components": {
-            "url": "/systems/" + str(project.system.id) + "/components/selected",
-            "title": "Manage Components",
-            "id": "project-components",
-        },
-        "ssp": {
-            "url": task.get_absolute_url(),
-            "title": "Export System Security Plan",
-            "id": "project-ssp"
-        },
-    }
-    return nav
 
 def project_edit(request, project_id):
     if request.method == 'POST':
